+++ packages/bootstrap-vue-3-icons/src/components/icons/helpers/make-icon.ts
@@ -1,7 +1,6 @@
-import { kebabCase, pascalCase } from '../../../utils/string'
+import {kebabCase, pascalCase} from '../../../utils/string'
 import BVIconBase from '../../BIcon/BIconBase.vue'
-import { defineComponent, h } from 'vue'
-
+import {defineComponent, h} from 'vue'
 
 /**
  * Icon component generator function
@@ -10,34 +9,26 @@ import { defineComponent, h } from 'vue'
  * @param {string} raw `innerHTML` for SVG
  * @return {VueComponent}
  */
-export const makeIcon = (name : string, content : string) => {
+export const makeIcon = (name: string, content: string) =>
   // For performance reason we pre-compute some values, so that
   // they are not computed on each render of the icon component
 
-  return defineComponent(
-    {
-      name: name,
-      extends : BVIconBase,
-      setup(props){
-        
-        const kebabName = kebabCase(name)
-        const iconName = `BIcon${pascalCase(name)}`
-        const iconNameClass = `bi-${kebabName}`
-        const iconTitle = kebabName.replace(/-/g, ' ')
-        const svgContent = (content || '').trim()
+  defineComponent({
+    name,
+    extends: BVIconBase,
+    setup() {
+      const kebabName = kebabCase(name)
+      const iconName = `BIcon${pascalCase(name)}`
+      const iconNameClass = `bi-${kebabName}`
+      const iconTitle = kebabName.replace(/-/g, ' ')
+      const svgContent = (content || '').trim()
 
-        return () =>{
-          return h(BVIconBase,
-           {title : iconTitle,
-            'aria-label': iconTitle ,
-            class: [iconNameClass],
-            content: svgContent
-          },
-        )
-        }
-       
-      }
-    }
-  )
-
-}
+      return () =>
+        h(BVIconBase, {
+          'title': iconTitle,
+          'aria-label': iconTitle,
+          'class': [iconNameClass],
+          'content': svgContent,
+        })
+    },
+  })