+++ packages/bootstrap-vue-3-icons/src/utils/string.ts
@@ -2,9 +2,7 @@ export const RX_HYPHENATE = /\B([A-Z])/g
 export const RX_UN_KEBAB = /-(\w)/g
 
 // Converts PascalCase or camelCase to kebab-case
-export const kebabCase = (str : string) => {
-  return str.replace(RX_HYPHENATE, '-$1').toLowerCase()
-}
+export const kebabCase = (str: string) => str.replace(RX_HYPHENATE, '-$1').toLowerCase()
 
 // Converts a kebab-case or camelCase string to PascalCase
 export const pascalCase = (str: string) => {