+++ packages/bootstrap-vue-3/src/components/BNavbar/BNavbarNav.vue
@@ -7,10 +7,10 @@
 <script setup lang="ts">
 import type {Alignment, Booleanish} from '../../types'
 import {computed, toRef} from 'vue'
-import {useBooleanish} from '../../composables'
+import {useAlignment, useBooleanish} from '../../composables'
 
 interface Props {
-  align?: Alignment
+  align?: Alignment.JustifyContent
   fill?: Booleanish
   justified?: Booleanish
   small?: Booleanish
@@ -27,11 +27,12 @@ const props = withDefaults(defineProps<Props>(), {
 const fillBoolean = useBooleanish(toRef(props, 'fill'))
 const justifiedBoolean = useBooleanish(toRef(props, 'justified'))
 const smallBoolean = useBooleanish(toRef(props, 'small'))
+const alignment = useAlignment(toRef(props, 'align'))
 
 const computedClasses = computed(() => ({
   'nav-fill': fillBoolean.value,
   'nav-justified': justifiedBoolean.value,
-  [`justify-content-${props.align}`]: props.align !== undefined,
+  [alignment.value]: props.align !== undefined,
   'small': smallBoolean.value,
 }))
 </script>