+++ packages/bootstrap-vue-3/src/components/BRow.vue
@@ -8,7 +8,7 @@
 import {computed, defineComponent, PropType, toRef} from 'vue'
 import {getBreakpointProps, getClasses} from '../utils'
 import type {Alignment, Booleanish} from '../types'
-import {useBooleanish} from '../composables'
+import {useAlignment, useBooleanish} from '../composables'
 
 const rowColsProps = getBreakpointProps('cols', [''], {type: [String, Number], default: null})
 
@@ -20,12 +20,13 @@ export default defineComponent({
     gutterY: {type: String, default: null},
     noGutters: {type: [Boolean, String] as PropType<Booleanish>, default: false},
     alignV: {type: String as PropType<Alignment.Vertical>, default: null},
-    alignH: {type: String as PropType<Alignment.Horizontal>, default: null},
+    alignH: {type: String as PropType<Alignment.JustifyContent>, default: null},
     alignContent: {type: String as PropType<Alignment.Content>, default: null},
     ...rowColsProps,
   },
   setup(props) {
     const noGuttersBoolean = useBooleanish(toRef(props, 'noGutters'))
+    const alignment = useAlignment(toRef(props, 'alignH'))
 
     const rowColsClasses = computed(() => getClasses(props, rowColsProps, 'cols', 'row-cols'))
 
@@ -36,7 +37,7 @@ export default defineComponent({
         [`gy-${props.gutterY}`]: props.gutterY !== null,
         'g-0': noGuttersBoolean.value,
         [`align-items-${props.alignV}`]: props.alignV !== null,
-        [`justify-content-${props.alignH}`]: props.alignH !== null,
+        [alignment.value]: props.alignH !== null,
         [`align-content-${props.alignContent}`]: props.alignContent !== null,
       },
     ])