+++ packages/bootstrap-vue-3/src/types/Size.d.ts
@@ -1,6 +1,11 @@
 /**
  * @external
  */
-type Size = 'sm' | 'lg' | undefined
+ export interface BaseSize {
+  'sm': unknown,
+  'lg': unknown,
+}
+
+type Size = keyof BaseSize | undefined
 
 export default Size